Load file methods
Shifted read_file() and get_line_data() to FileModule.FilerModule.py
